requirements.txt
Flask == 2.0.3
PyJWT == 2.3.0
PyMySQL == 0.10.1
SQLAlchemy == 1.4.18
SQLAlchemy-Utils == 0.38.2




